requirements.txt
appnope==0.1.0
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.2
dialogflow==0.7.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
google-api-core==1.16.0
google-auth==1.11.2
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.4.0
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.1.0
packaging==20.1
parso==0.6.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
pyparsing==2.4.6
pytest==5.3.4
pytest-mock==2.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
soupsieve==2.0
traitlets==4.3.3
urllib3==1.25.8
virtualenv==16.7.7
wcwidth==0.1.8
Werkzeug==0.16.0
zipp==2.0.1
